Move renderer.h, ViewMatrix to DLE.Renderer
