+++ DLE/Editor/DLE.Editor.vcxproj.filters
@@ -49,9 +49,6 @@
     <ClInclude Include="mineview.h">
       <Filter>UI\Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="renderer.h">
-      <Filter>UI\Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="StdAfx.h">
       <Filter>UI\Header Files</Filter>
     </ClInclude>
@@ -71,9 +68,6 @@
       <Filter>UI\Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Global.h">
-      <Filter>UI\Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="ViewMatrix.h">
       <Filter>UI\Header Files</Filter>
     </ClInclude>
     <ClInclude Include="DrawHelpers.h">
@@ -280,9 +274,6 @@
       <Filter>UI\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="MineView.Input.cpp">
-      <Filter>UI\Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="ViewMatrix.cpp">
       <Filter>UI\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="global.cpp">
